requirements.txt
numpy
pandas
scikit-learn
matplotlib
joblib
pyyaml
pytest
